chat section made as component
